Adjusted the data structure of the returned object from fetchProductById and made product catalog re-render when a new product is added
